move `Cow[u8]` conversions into `conversions::std::slice` module
